requirements.txt
rpi-lgpio==0.6
certifi==2022.9.24
chardet==5.1.0
colorzero==2.0
distro==1.8.0
gpiozero==2.0.1
idna==3.3
olefile==0.46
pigpio==1.78
Pillow==9.4.0
pycryptodomex==3.11.0
requests==2.28.1
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
urllib3==1.26.12
